Added cookie formats supporting

+++ BotsCommon/TimeConverter.cs
@@ -4,7 +4,7 @@ namespace BotsCommon
 {
     public static class TimeConverter
     {
-        public static DateTime FromUnixTimestamp(long unixTime)
+        public static DateTime FromUnixTimestamp(double unixTime)
         {
             return new DateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeKind.Utc).AddSeconds(unixTime).ToLocalTime();
         }
